Skill expertise won't fraction

+++ lib/features/user_profile/views/widgets/professional_skill_list_item.dart
@@ -68,7 +68,7 @@ class ProfessionalSkillListItem extends StatelessWidget {
                 Text(
                   skillInfo.rating == null
                       ? "N/A"
-                      : skillInfo.rating.toString() + "/10",
+                      : skillInfo.rating.toInt().toString() + "/10",
                   style: titleStyle,
                 ),
                 if (isInEditMode)
